Import PDFViewer Component from Angular
In React, we now have access to a pdf-viewer component, which is a simple wrapper of the component from ngx-extended-pdf-viewer.

pdf-viewer takes in a src for the pdf file and emits a pdfService object, that can be used to access the pdf data.

In order to make any changes to the core pdf-viewer component, the angular project needs to be updated and re-exported (contact Chase for this).